fix cmt - change path api, move logic get to userservices
