+++ internal/app/api/services/user/user.go
@@ -2,6 +2,7 @@ package userservices
 
 import (
 	"context"
+	"strconv"
 	"time"
 
 	"dating/internal/app/api/types"
@@ -21,6 +22,8 @@ type Repository interface {
 	UpdateUserByID(ctx context.Context, User types.User) error
 	GetListUsers(ctx context.Context, ps types.PagingNSorting) ([]*types.UserResGetInfo, error)
 	CountUser(ctx context.Context) (int, error)
+	GetListlikedInfo(ctx context.Context, idUser string) ([]*types.UserResGetInfo, error)
+	GetListMatchedInfo(ctx context.Context, idUser string) ([]*types.UserResGetInfo, error)
 }
 
 // Service is an user service
@@ -191,4 +194,58 @@ func (s *Service) GetListUsers(ctx context.Context, page, size string) (*types.G
 	s.logger.Infof("get list users by page is completed, page: ", pagingNSorting)
 
 	return &listUsersResponse, nil
+}
+
+// get list user liked
+func (s *Service) listLiked(ctx context.Context, userID string) ([]types.UserResGetInfo, error) {
+	list, err := s.repo.GetListlikedInfo(ctx, userID)
+	if err != nil {
+		return nil, err
+	}
+	return s.convertPointerArrayToArray(list), err
+}
+
+// get list user matched
+func (s *Service) listMatched(ctx context.Context, userID string) ([]types.UserResGetInfo, error) {
+	list, err := s.repo.GetListMatchedInfo(ctx, userID)
+	if err != nil {
+		return nil, err
+	}
+	return s.convertPointerArrayToArray(list), err
+}
+
+// get list matched or liked
+func (s *Service) GetMatchedUsersByID(ctx context.Context, idUser, matchedParameter string) ([]types.UserResGetInfo, error) {
+
+	matched, err := strconv.ParseBool(matchedParameter)
+
+	if err != nil {
+		s.logger.Errorf("Failed url parameters when get list matched or like", err)
+		return nil, errors.Wrap(err, "Failed url parameters when get list users")
+	}
+
+	if !bson.IsObjectIdHex(idUser) {
+		s.logger.Errorf("Id user incorrect,it isn't ObjectIdHex")
+		return nil, errors.New("Id user incorrect to find list liked from database, it isn't ObjectIdHex")
+	}
+
+	if matched {
+		list, err := s.listMatched(ctx, idUser)
+		s.logger.Infof("Get list matched completed", idUser)
+		return list, err
+	}
+
+	list, err := s.listLiked(ctx, idUser)
+	s.logger.Infof("Get list liked completed", idUser)
+	return list, err
+}
+
+// convert []*types.UserResGetInfo to []types.UserResGetInfo - if empty return []
+func (s *Service) convertPointerArrayToArray(list []*types.UserResGetInfo) []types.UserResGetInfo {
+
+	listUsers := []types.UserResGetInfo{}
+	for _, user := range list {
+		listUsers = append(listUsers, *user)
+	}
+	return listUsers
 }
